requirements.txt
yfinance==0.2.40
pandas==2.2.2
pandas-ta==0.3.14b
numpy==1.26.4
scipy==1.13.0
scikit-learn==1.4.1
requests==2.31.0
fredapi==0.5.2
finnhub-python==2.4.20
sec-edgar-api==1.0.6
lxml==5.2.1
html5lib==1.1
